libntfs: Tidy ups and minor tweaks

+++ branches/newgui/libs/libntfs/source/ntfsfile.h
@@ -42,8 +42,8 @@ typedef struct _ntfs_file_state {
     bool append;                            /* True if allowed to append to file */
     bool compressed;                        /* True if file data is compressed */
     bool encrypted;                         /* True if file data is encryted */
-    s64 pos;                                /* Current position within the file (in bytes) */
-    s64 len;                                /* Total length of the file (in bytes) */
+    off_t pos;                              /* Current position within the file (in bytes) */
+    size_t len;                             /* Total length of the file (in bytes) */
     struct _ntfs_file_state *prevOpenFile;  /* The previous entry in a double-linked FILO list of open files */
     struct _ntfs_file_state *nextOpenFile;  /* The next entry in a double-linked FILO list of open files */
 } ntfs_file_state;
